+++ DB/src/test/java/io/deephaven/db/tables/QueryTableHugeSortTest.java
@@ -5,11 +5,8 @@ import io.deephaven.db.tables.select.QueryScope;
 import io.deephaven.db.tables.utils.TableTools;
 import io.deephaven.db.v2.sources.AbstractColumnSource;
 import io.deephaven.db.v2.utils.Index;
-import io.deephaven.test.junit4.DbTestRule;
+import io.deephaven.test.junit4.DbCleanup;
 import io.deephaven.test.types.OutOfBandTest;
-import io.deephaven.util.ExceptionDetails;
-import junit.framework.TestCase;
-import org.junit.After;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -110,5 +107,5 @@ public class QueryTableHugeSortTest {
     }
 
     @Rule
-    public final DbTestRule rule = new DbTestRule();
+    public final DbCleanup rule = new DbCleanup();
 }
