+++ DB/src/test/java/io/deephaven/db/tables/verify/TestTableAssertions.java
@@ -5,7 +5,7 @@ import io.deephaven.db.tables.Table;
 import io.deephaven.db.tables.live.LiveTableMonitor;
 import io.deephaven.db.tables.utils.TableTools;
 import io.deephaven.db.v2.*;
-import io.deephaven.test.junit4.JUnit4LiveTableTestCase;
+import io.deephaven.test.junit4.DbCleanup;
 import io.deephaven.util.QueryConstants;
 import junit.framework.TestCase;
 import org.junit.Rule;
@@ -19,7 +19,7 @@ import static io.deephaven.db.v2.TstUtils.*;
 
 public class TestTableAssertions {
     @Rule
-    public final JUnit4LiveTableTestCase base = new JUnit4LiveTableTestCase();
+    public final DbCleanup base = new DbCleanup();
 
     @Test
     public void testStatic() {
