+++ DB/src/test/java/io/deephaven/db/v2/select/TestFormulaColumn.java
@@ -15,7 +15,7 @@ import io.deephaven.db.tables.lang.DBLanguageParser.QueryLanguageParseException;
 import io.deephaven.db.tables.libs.QueryLibrary;
 import io.deephaven.db.tables.select.QueryScope;
 import io.deephaven.db.tables.utils.*;
-import io.deephaven.test.junit4.JUnit4QueryTableTestBase;
+import io.deephaven.test.junit4.DbCleanup;
 import io.deephaven.db.v2.utils.codegen.TypeAnalyzer;
 import io.deephaven.test.types.OutOfBandTest;
 import io.deephaven.util.QueryConstants;
@@ -68,7 +68,7 @@ public class TestFormulaColumn {
     }
 
     @Rule
-    public final JUnit4QueryTableTestBase base = new JUnit4QueryTableTestBase();
+    public final DbCleanup base = new DbCleanup();
 
     @Before
     public void setUp() throws Exception {
